backend/requirements.txt
fastapi
uvicorn[standard]
kerykeion
geopy
pytz
timezonefinder
google-generativeai
pydantic
python-dotenv
httpx
